+++ backend/backend/src/main/java/com/dypaworld/config/CorsConfig.java
@@ -18,7 +18,8 @@ public class CorsConfig {
                         .allowedOrigins("http://localhost:5173")
                         .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
                         .allowedHeaders("*")
-                        .allowCredentials(true);
+                        .allowCredentials(true)
+                        .exposedHeaders("*");
             }
         };
     }
