+++ backend/backend/src/main/java/com/dypaworld/service/CustomOAuth2UserService.java
@@ -18,7 +18,7 @@ public class CustomOAuth2UserService implements OAuth2UserService<OAuth2UserRequ
 
     @Override
     public OAuth2User loadUser(OAuth2UserRequest userRequest) throws OAuth2AuthenticationException {
-        System.out.println("Madisong give me the life support I need");
+        System.out.println("Madison give me the life support I need");
         OAuth2User oAuth2User = new DefaultOAuth2UserService().loadUser(userRequest);
 
         System.out.println("ATTRIBUTI: " + oAuth2User.getAttributes());
