+++ backend/backend/src/main/resources/application.properties
@@ -16,7 +16,7 @@ spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL8Dialect
 
 # === Security Settings ===
 spring.security.user.name=${SPRING_SECURITY_USER}
-spring.security.user.password=${SPRING_SECURITY_PASSWORD}
+#spring.security.user.password=${SPRING_SECURITY_PASSWORD}
 
 # === Github OAuth2 Settings ===
 spring.security.oauth2.client.registration.github.client-id=${GITHUB_CLIENT_ID}
